[refactor] Make 'pylint.testutil.primer' a private API
