+++ doc/whatsnew/2/2.15/index.rst
@@ -47,3 +47,7 @@ Other Changes
 
 Internal changes
 ================
+
+* ``pylint.testutils.primer`` is now a private API.
+
+  Refs #6905
