+++ pylint/testutils/_primer/package_to_lint.py
